added floating npc fix, and fixed price variable issues
added npc floating fix, updated sellprice, and buyprice to be item_price, so there is less chance of mistakes.added npc floating fix, updated sellprice, and buyprice to be item_price, so there is less chance of mistakes. wrong confg sell or buy variable throws errors, this reduces the change of using the wrong variable.